auth-service/requirements.txt
django==5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.11
redis==5.0.1
django-redis==5.4.0
django-cors-headers==4.3.1
gunicorn==21.2.0
python-decouple==3.8